+++ backend/windmill-common/src/error.rs
@@ -96,9 +96,9 @@ impl IntoResponse for Error {
         };
 
         if matches!(status, axum::http::StatusCode::NOT_FOUND) {
-            tracing::warn!(not_found = e.to_string());
+            tracing::warn!(message = e.to_string());
         } else {
-            tracing::error!(error = e.to_string());
+            tracing::error!(message = e.to_string());
         };
 
         axum::response::Response::builder()