+++ frontend/src/lib/components/DisplayResult.svelte
@@ -359,7 +359,7 @@
 				{/each}
 			</div>
 		{:else if largeObject}
-			{#if 'filename' in result && 'file' in result}
+			{#if typeof result == 'object' && 'filename' in result && 'file' in result}
 				<div
 					><a
 						download={result.filename ?? result.file?.filename ?? 'windmill.file'}