+++ frontend/src/lib/components/apps/editor/settingsPanel/GridTabDisabled.svelte
@@ -19,9 +19,10 @@
 	on:change={() => {
 		if (disablable) {
 			field = {
-				type: 'eval',
+				type: 'evalv2',
 				expr: 'false',
-				fieldType: 'boolean'
+				fieldType: 'boolean',
+				connections: []
 			}
 		} else {
 			field = {